+++ HISTORY.rst
@@ -2,6 +2,8 @@
 History
 =======
 
+* Add Django 2.0+ fixer to rewrite imports of ``lru_cache`` from ``django.utils.functional`` to use ``functools``.
+
 1.3.2 (2021-09-23)
 ------------------